Moved and splitted moderation commands
